Split prototype into smaller apps.
The Django way is to break things up into smaller apps. I've moved the
code for displaying organisations into the organisations app and the
models for questions and answers into the questions app.

Organisation authentication is working. After logging in the
Organisation is redirected to the list of current questions.